More naming convention fixes.
Fixed thingTranslator naming conventions to match what opent2t library expects, and updated RAML conventions to map to function names in thingTranslator.

Test automation is currently failing, but due to an actual issue with the wink translator! So the test is actually working, but the translator is legitmately failing. That will be addressed separately as part of #52.