Fixed wink api failure #52.
As part of this, I have moved the wink helper to the v2 api and signficantly cleaned it up. I have fixed the wink thermostat translator, but now the other wink translators need to be fixed as well. I will follow up with subsequent checkins to resolve as part of #40.